+++ lib/tracer/iex.ex
@@ -1,4 +1,2 @@
 defmodule IEx.Helpers.Tracer do
-
 end
-